Upgraded the relatedWorks serivce to align with SOA
