+++ backend/services/relatedWorks/controllers/relatedWorksController.js
@@ -1,0 +1,44 @@
+import { getRelatedWorks } from "../service/RelatedWorksService.js";
+import { db } from "../firebaseConfig.js";
+
+
+// Creates the related works from the relatedWorksService 
+export const relatedWorks = async (req, res) => {
+    try {
+        const userId = req.user.uid; 
+        const { fileName } = req.body;
+        if (!fileName || !userId) return res.status(400).json({ error: "fileName and userId are required" });
+
+        const finalData = await getRelatedWorks(fileName, userId );
+        res.json(finalData);
+    } catch (error) {
+        console.error("Error generating related works:", error.message);
+        res.status(500).json({ error: error.message || "Internal Server Error" });
+    }
+}
+
+// Retrieves the related works from the realtime Firebase db
+export const getRelatedWorks = async (req, res) => {
+    try {
+        const userId = req.user.uid; 
+        const { fileName } = req.query;
+
+        if (!fileName || !userId) {
+            return res.status(400).json({ error: "fileName and userId are required"});
+
+        }
+
+        const snapshot = await db.ref(`users/${userId}/data/${fileName}/related`).once("value");
+        const storedData = snapshot.val();
+
+        if (!storedData) {
+            return res.status(404).json(null);
+        }
+
+        res.json(storedData);
+    } catch (error) {
+        console.error("Error fetching stored data:", error);
+        res.status(500).json({error: "Failed to retrieve stored data"})
+    }
+
+}
